+++ build.common.gradle
@@ -148,7 +148,9 @@ dependencies {
     implementation 'org.ftclib.ftclib:core:2.1.1'
     // the following is optional if you want vision
     implementation 'org.ftclib.ftclib:vision:2.1.0'
-    implementation 'org.openftc:apriltag:2.0.0'
+
+//    add back when swerve is working
+//    implementation 'org.openftc:apriltag:2.0.0'
 
 }
